성공: Constraint DSL 연결 성공
