+++ Sources/SwiftLayout/SwiftLayout.swift
@@ -9,18 +9,3 @@ import Foundation
 import UIKit
 
 public struct SwiftLayout {}
-
-public protocol Layoutable: CustomDebugStringConvertible {
-    
-    var branches: [Layoutable] { get }
-    
-    func isEqualLayout(_ layoutable: Layoutable) -> Bool
-    func isEqualView(_ layoutable: Layoutable) -> Bool
-    func isEqualView(_ view: UIView?) -> Bool
-    
-    var layoutIdentifier: String { get }
-}
-
-extension Layoutable {
-    public var branches: [Layoutable] { [] }
-}
